requirements.txt
boto3~=1.17.101
python-dotenv~=0.18.0